Updating files for ACI timing experiment. Should be right this time - I think the problem was in the slurm.sh submit script. There still could be a problem, but I might get lucky and be able to get around it in the submit script.
The potential problem is that all 16 corse on each node I take need to be allocated to my job (so that others cant use them), but I want MPI to only run on one of them so that OpenMP can work. I think the submit script will allow for this now.